Move the test resource out of `./webhook`.
This is so that other things can more naturally write tests using this same test resource.
